Refactor and reorganize codebase
- Removed the main application entry point file `pyrox/__main__.py`.
- Refactored `SliceableInt` class in `pyrox/models/abc/meta.py` for improved readability.
- Added unit tests for bit operations in `pyrox/models/abc/test/test_meta.py`.
- Updated import paths in `pyrox/models/plc/controller.py` and `pyrox/services/byte.py`.
- Removed unused code and files related to fuzzy operations and tasks in `pyrox/tasks`.
- Introduced new `bit.py` service for bit manipulation functions.
- Created a new `search.py` service for utility functions related to pattern matching and duplicates.
- Added tests for bit operations in `pyrox/services/test/test_bit.py`.
