+++ pyrox/tasks/__init__.py
@@ -1,6 +1,5 @@
-from . import builtin, tools
+from . import builtin
 
 __all__ = (
     'builtin',
-    'tools',
 )
