--- pyrox/tasks/ladder.py
@@ -1,49 +1,0 @@
-"""PLC Ladder Editor application task module.
-"""
-from __future__ import annotations
-import importlib
-
-from pyrox.applications.app import App, AppTask
-from pyrox.models.gui import ladder as ladder_gui
-
-from pyrox.models.plc import Routine
-from tkinter import Menu
-
-
-class RoutineTask(AppTask):
-    """Routine task for generating and managing PLC routines.
-    """
-
-    def __init__(self,
-                 application: App):
-        super().__init__(application=application)
-
-    def debug_draw(self):
-        """Debug method to draw the ladder editor frame."""
-        self.logger.info('Drawing ladder editor frame for debugging...')
-        importlib.reload(ladder_gui)
-        ladder_frame = ladder_gui.LadderEditorTaskFrame(
-            master=self.application.workspace)
-        self.application.register_frame(ladder_frame, raise_=True)
-
-    def generate_default_routine(self):
-        """Generate a default routine for the PLC application."""
-        self.logger.info('Generating default routine...')
-        if not self.application.controller:
-            self.logger.error('No controller set in the application.')
-            return
-        importlib.reload(ladder_gui)
-        routine = Routine(name='DefaultRoutine',
-                          description='Generated by Pyrox Programming Application.')
-        ladder_frame = ladder_gui.LadderEditorTaskFrame(
-            master=self.application.workspace,
-            controller=self.application.controller,
-            routine=routine)
-        self.application.register_frame(ladder_frame, raise_=True)
-
-    def inject(self) -> None:
-        drop_down = Menu(self.application.menu.tools, name='generate_routine', tearoff=0)
-        self.application.menu.tools.insert_cascade(0, label='Generate Routine', menu=drop_down)
-
-        drop_down.add_command(label='Generate Default Routine', command=self.generate_default_routine)
-        drop_down.add_command(label='Draw', command=self.debug_draw)
